Switch to Vertico minibuffer completion package
Replace Selectrum with Vertico. It's not clear cut which one is better.
I like the Vertico completion with few input characters better as it
prefers to match file prefixes as long as there are matches.

Supposedly TRAMP file access may be slow with Vertico. I also had
trouble to enable the mouse selection extension with the straight
package manager. I rarely use the mouse so can live with that.

Both Selectrum and Vertico are essentially one-man shows so it's
difficult to know which one (if any) of them is the best long term
choice.