redefine key bindings for next and previous error
Use M-] and C-M-] for next and previous error respectively. These works
better for HHKB keyboards without function keys. The previous bindings
was F9 and F10 (and there is a default binding for F10 which was
overridden).

Unfortunately M-[ can't be used in terminals, otherwise it would pair
better with M-].